Add game over and title screen

+++ matrix.h
@@ -6,14 +6,14 @@
 #include "stm32f0_discovery.h"
 
 void generate_image(void);
-void generate_row(int);
-void mask_canvas(int, int, int, int, uint8_t(*)[]);
+void generate_row(short);
+void mask_canvas(short, short, short, short, uint8_t(*)[]);
 void clear_display();
 void gameover();
 void draw_title();
 void draw_background();
-void draw_bird(int, short);
-void draw_pipe(int, int);
+void draw_bird(short, short);
+void draw_pipe(short, short);
 void bird_fly();
 //void clear_pipe(double, double, double);
